Add gdscript support (upgrade tree-sitter-wasms to 0.4.0) (#112)
* Start gdscript support

* Fix parser loc snafu; increment plugin ver

For some inscrutable reason, the upgrade to tree-sitter-wasms 0.4.0 causes this project's buildscripts to put the parsers in `parsers/out/` instead of just `parsers/`. The root path to where the parsers are located has been changed in `src/extension.ts` to account for this... it could probably change again, though.

yarn pls

* Update whitespace as suggested

Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>

* Update extension.ts

Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>

---------

Co-authored-by: Andreas Arvidsson <andreas.arvidsson87@gmail.com>